general updates

+++ otashu-android/src/com/andrewsummers/otashu/activity/ApprenticeActivity.java
@@ -147,7 +147,7 @@ public class ApprenticeActivity extends Activity implements OnClickListener {
 
         // fill bundle with values we are passing to next activity
         Bundle bundle = new Bundle();
-        bundle.putInt("emotion_focus_id", selectedEmotionFocusValue);
+        bundle.putLong("emotion_focus_id", selectedEmotionFocusValue);
 
         switch (v.getId()) {
             case R.id.button_choose_apprentice:
